phase-54: HS-54-02 behavior modules — the 2,907-line monolith becomes twelve single-concern ES modules
dictation-app.js is now a 19-line entry; the behavior lives under
scripts/dictation/ (core, blocks, readiness, knowledge, runtime, memory,
journal, dryrun, agent, discovery-nudge, activity-nudges, init — largest
576 lines, all under the ~600 budget). Code moved verbatim; the one
structural idiom is core's section-loader registry
(registerSection/loadSection), which keeps the module graph acyclic
where direct imports would have cycled (knowledge<->readiness,
core<->features). Three shared HTML helpers (renderRuntimeGuidance,
renderDryTelemetry, learnSigChip) moved to core for the same reason.
The HS-54-01 dependency injection in discovery-nudge.js is gone — it
imports api/projectRootParam from core directly, as planned.

Zero test files changed. Slice: 158 passed. Full suite: 2540 passed,
17 skipped. npm run build clean; 0 _built/ tracked. Proven live by a
new all-tabs Playwright dogfood (dogfood_story02.py): nine tabs
activate + populate, pipeline enabled via the UI, dry-run renders final
text + stage trace + moment-of-truth, the ritual acknowledges, the run
journals, a correction adds + deletes — 16/16 PASS with zero uncaught
page errors. Two early dogfood failures were wrong assumptions (a
disabled pipeline's dry-run never journals — verified by probing the
API), not carve regressions.
